refactor: split profile step page
